+++ STACpopulator/implementations/HRDPS_CRIM/__init__.py
@@ -1,0 +1,3 @@
+from .add_HRDPS import HRDPSpopulator as Populator
+
+__all__ = ["Populator"]