+++ STACpopulator/implementations/RDPS_CRIM/__init__.py
@@ -1,0 +1,3 @@
+from .add_RDPS import RDPSpopulator as Populator
+
+__all__ = ["Populator"]