+++ STACpopulator/populators.py
@@ -1,12 +1,15 @@
 import argparse
 import functools
+import importlib
+import importlib.util
 import inspect
 import json
 import logging
 import os
+import re
 from abc import ABC, abstractmethod
 from datetime import datetime
-from typing import Any, Dict, Iterable, List, MutableMapping, Optional, Type, Union, get_args
+from typing import Any, Callable, Dict, Iterable, List, MutableMapping, Optional, Type, Union, get_args
 
 import pystac
 import requests
@@ -19,6 +22,7 @@ from STACpopulator.api_requests import (
     stac_version_match,
 )
 from STACpopulator.collection_update import UpdateModesOptional, update_collection
+from STACpopulator.exceptions import FunctionLoadError
 from STACpopulator.input import ErrorLoader, GenericLoader, THREDDSLoader
 from STACpopulator.models import AnyGeometry
 from STACpopulator.stac_utils import load_config
@@ -36,6 +40,9 @@ class STACpopulatorBase(ABC):
         update: bool = False,
         session: Optional[Session] = None,
         config_file: Optional[Union[os.PathLike[str], str]] = "collection_config.yml",
+        extra_item_parsers: Optional[list[str]] = None,
+        extra_collection_parsers: Optional[list[str]] = None,
+        extra_parser_arguments: Optional[dict[str, str] | list[tuple[str, str]]] = None,
         update_collection: UpdateModesOptional = "none",
         exclude_summaries: Iterable[str] = (),
     ) -> None:
@@ -50,6 +57,19 @@ class STACpopulatorBase(ABC):
         self._collection_config_path = config_file
         self._collection_info: MutableMapping[str, Any] = None
         self._session = session
+        extra_parser_arguments = dict(extra_parser_arguments or {})
+        self._extra_item_parsers = [
+            self._load_extra_parser(parser, extra_parser_arguments) for parser in (extra_item_parsers or [])
+        ]
+        self._extra_collection_parsers = [
+            self._load_extra_parser(parser, extra_parser_arguments) for parser in (extra_collection_parsers or [])
+        ]
+
+        if extra_parser_arguments and not (self._extra_collection_parsers or self._extra_item_parsers):
+            LOGGER.warning(
+                "extra_parser_arguments will be ignored because no extra collection or item parsers are specified."
+            )
+
         self.load_config()
 
         self._ingest_pipeline = data_loader
@@ -61,6 +81,37 @@ class STACpopulatorBase(ABC):
         LOGGER.info("Initialization complete")
         LOGGER.info(f"Collection {self.collection_name} is assigned ID {self.collection_id}")
         self._collection = self.create_stac_collection()
+
+    @staticmethod
+    def _load_extra_parser(func_str: str, extra_kwargs: dict[str, str]) -> Callable:
+        if ":" in func_str:
+            mod, func = func_str.split(":", 1)
+            if mod.endswith(".py"):
+                mod_name = re.sub(r"\W", "_", os.path.splitext(os.path.basename(mod))[0])
+                mod_spec = importlib.util.spec_from_file_location(mod_name, mod)
+                function_ns = importlib.util.module_from_spec(mod_spec)
+                try:
+                    mod_spec.loader.exec_module(function_ns)
+                except FileNotFoundError as e:
+                    raise FunctionLoadError(f"Unable to load python module from file: '{mod}'") from e
+            else:
+                try:
+                    function_ns = importlib.import_module(mod)
+                except ModuleNotFoundError as e:
+                    raise FunctionLoadError(f"Unable to load module '{mod}'") from e
+            try:
+                callable_func = getattr(function_ns, func)
+            except AttributeError as e:
+                raise FunctionLoadError(f"Unable to load function '{func}' from '{mod}'") from e
+            arg_spec = inspect.getfullargspec(callable_func)
+            if arg_spec.varkw:
+                return functools.partial(callable_func, **extra_kwargs)
+            all_kwargs = arg_spec.args + arg_spec.kwonlyargs
+            return functools.partial(callable_func, **{k: v for k, v in extra_kwargs.items() if k in all_kwargs})
+        else:
+            raise FunctionLoadError(
+                "Parser function string is not properly formatted. Should be in the form 'module:function_name'"
+            )
 
     def load_config(self) -> None:
         """
@@ -146,6 +197,14 @@ class STACpopulatorBase(ABC):
         # Add any assets if provided in the config
         self._collection_info["assets"] = self.__make_collection_assets()
 
+        # Cast providers to pystac objects
+        self._collection_info["providers"] = self.__make_collection_providers()
+
+        # Add contacts as extra_field
+        if "contacts" in self._collection_info:
+            self._collection_info["extra_fields"] = {}
+            self._collection_info["extra_fields"]["contacts"] = self.__make_collection_contacts()
+
         # Construct links if provided in the config. This needs to be done before constructing a collection object.
         collection_links = self.__make_collection_links()
 
@@ -155,6 +214,8 @@ class STACpopulatorBase(ABC):
             collection.add_links(collection_links)
         collection.add_links(self._ingest_pipeline.links)
         collection_data = collection.to_dict()
+        for func in self._extra_collection_parsers:
+            func(collection_data)
         self.publish_stac_collection(collection_data)
         return collection_data
 
@@ -182,6 +243,29 @@ class STACpopulatorBase(ABC):
                 pystac_assets[asset_name] = pystac.Asset(**asset_info)
         return pystac_assets
 
+    def __make_collection_providers(self) -> List[pystac.Provider]:
+        """Create collection level providers based on data read in from the configuration file.
+
+        :return: List of pystac Provider objects
+        :rtype: List[pystac.Provider]
+        """
+        pystac_providers = []
+        if "providers" in self._collection_info:
+            providers = self._collection_info.pop("providers")
+            pystac_providers = [pystac.Provider(**provider) for provider in providers]
+        return pystac_providers
+
+    def __make_collection_contacts(self) -> List[dict]:
+        """Create collection level contacts based on data read in from the configuration file.
+
+        :return: List of dictionnary contact objects
+        :rtype: List[dict]
+        """
+        contacts = []
+        if "contacts" in self._collection_info:
+            contacts = self._collection_info.pop("contacts")
+        return contacts
+
     def publish_stac_collection(self, collection_data: dict[str, Any]) -> None:
         """Publish this collection by uploading it to the STAC catalog at self.stac_host."""
         post_stac_collection(self.stac_host, collection_data, self.update, session=self._session)
@@ -199,6 +283,8 @@ class STACpopulatorBase(ABC):
             LOGGER.info(f"New data item: {item_name}", extra={"item_loc": item_loc})
             try:
                 stac_item = self.create_stac_item(item_name, item_data)
+                for func in self._extra_item_parsers:
+                    func(stac_item)
             except Exception:
                 LOGGER.exception(
                     f"Failed to create STAC item for {item_name}",
@@ -238,6 +324,12 @@ class STACpopulatorBase(ABC):
         if self.update and self.update_collection != "none":
             self.publish_stac_collection(self._collection)
 
+    @staticmethod
+    def _extra_parser_argument(arg: str) -> tuple[str, str]:
+        if "=" in arg:
+            return tuple(a.strip() for a in arg.split("=", 1))
+        raise argparse.ArgumentTypeError("--extra-parser-arguments must be in the form 'key=value'")
+
     @classmethod
     def update_parser_args(cls, parser: argparse.ArgumentParser) -> None:
         """Add additional CLI arguments to the argument parser."""
@@ -257,6 +349,34 @@ class STACpopulatorBase(ABC):
             action="extend",
             default=[],
             help="Exclude these properties when updating collection summaries. ",
+        )
+        parser.add_argument(
+            "-x",
+            "--extra-item-parsers",
+            action="append",
+            help="Functions that may modify items before upload. "
+            "Should be specified in the form 'module:function_name' "
+            "and have the signature function(item: dict, **kw)",
+        )
+        parser.add_argument(
+            "-X",
+            "--extra-collection-parsers",
+            action="append",
+            help="Functions that may modify collections before upload. "
+            "Should be specified in the form 'module:function_name' or "
+            "path/to/python/file.py:function_name. Functions should "
+            "have the signature function(collection: dict, **kw) -> None "
+            "and should modify the collection dict in place.",
+        )
+        parser.add_argument(
+            "-a",
+            "--extra-parser-arguments",
+            action="append",
+            type=cls._extra_parser_argument,
+            help="Extra keyword arguments that should be passed to extra "
+            "item and collection function as "
+            "keyword arguments. "
+            "Should be specified in the form 'key=value'",
         )
 
     @classmethod